Research/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aiosqlite==0.20.0
python-multipart==0.0.12
httpx==0.27.2
openai==1.54.0
PyMuPDF==1.24.13
python-docx==1.1.2
bibtexparser==1.4.2
textstat==0.7.3
arxiv==2.1.3
websockets==13.1
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2
